split run from cron
