+++ .env.example
@@ -1,3 +1,7 @@
+# Mentor Config
+RUN_EXPORT_TO_SHEETS_EVERY_MINUTES=30
+RUN_INTEGRATIONS_EVERY_MINUTES=30
+
 # CYFBot Config
 ACCESS_CHANNEL_ID="Cxxxxxxxxxx"
