+++ package.json
@@ -10,8 +10,11 @@
     "build": "tsc",
     "postinstall": "npm run build",
     "start": "node build/index.js",
-    "dev": "PORT=5000 nodemon --exec ts-node src/index.ts",
-    "data": "ts-node src/cron.ts"
+    "talk-to-bot": "PORT=5000 nodemon --exec ts-node src/services/talk-to-bot/index.ts",
+    "integrations:run": "ts-node src/services/integrations/index.ts",
+    "integrations:cron": "ts-node src/services/integrations/cron.ts",
+    "export-to-sheets:run": "ts-node src/services/export-to-sheets/index.ts",
+    "export-to-sheets:cron": "ts-node src/services/export-to-sheets/cron.ts"
   },
   "dependencies": {
     "@slack/bolt": "^3.6.0",
