+++ src/services/integrations/cron.ts
@@ -1,0 +1,11 @@
+import { integrations } from ".";
+
+var cron = require("node-cron");
+
+(async () => {
+  console.log("fetch integrations");
+  await integrations();
+  cron.schedule("*/30 * * * *", () => {
+    integrations();
+  });
+})();
