+++ src/services/talk-to-bot/listeners/clickDataDump.ts
@@ -1,5 +1,5 @@
 import fieldModal from "../blocks/fieldModal";
-import { getSchema } from "../clients/apollo";
+import { getSchema } from "../../../clients/apollo";
 
 export default function (slack) {
   slack.action("CLICK_DATA_DUMP", async ({ ack, body, client }: any) => {
