+++ src/services/talk-to-bot/listeners/clickEditSchema.ts
@@ -1,6 +1,6 @@
 import fieldModal from "../blocks/fieldModal";
 import schemaModal from "../blocks/schemaModal";
-import { getSchema } from "../clients/apollo";
+import { getSchema } from "../../../clients/apollo";
 
 // When the user clicks a button to edit a value
 // open the modal to edit the value.
