+++ src/services/talk-to-bot/listeners/clickExportData.ts
@@ -1,6 +1,6 @@
 import fieldModal from "../blocks/fieldModal";
-import { getSchema } from "../clients/apollo";
-import sheets from "../sheets";
+import { getSchema } from "../../../clients/apollo";
+import { sheets } from "../../export-to-sheets";
 
 // When the user clicks a button to edit a value
 // open the modal to edit the value.
