+++ src/services/talk-to-bot/listeners/mentionChannel.ts
@@ -1,4 +1,4 @@
-import database, { getSchema } from "../clients/apollo";
+import database, { getSchema } from "../../../clients/apollo";
 import getStudent from "../queries/getStudent.graphql";
 import getCheckInReporters from "../queries/getCheckInReporters.graphql";
 import { json2csvAsync } from "json-2-csv";
