+++ src/services/talk-to-bot/listeners/mentionStudent.ts
@@ -1,7 +1,7 @@
-import fieldsList from '../blocks/fieldsList';
-import quickUpdateButtons from '../blocks/quickUpdateButtons';
-import database, { getSchema } from '../clients/apollo';
-import getStudent from '../queries/getStudent.graphql';
+import fieldsList from "../blocks/fieldsList";
+import quickUpdateButtons from "../blocks/quickUpdateButtons";
+import database, { getSchema } from "../../../clients/apollo";
+import getStudent from "../queries/getStudent.graphql";
 
 // When the user @mention's a student
 // Show them the student home form
@@ -16,12 +16,12 @@ export const studentHome = async (params: StudentHome) => {
   const { data } = await database.query({
     query: getStudent,
     variables: { studentID: params.studentID },
-    fetchPolicy: 'network-only',
+    fetchPolicy: "network-only",
   });
   const schema = await getSchema();
   const quickButtons = quickUpdateButtons(params);
   const fieldButtons = fieldsList({ ...params, data, schema });
-  return [...quickButtons, { type: 'divider' }, ...fieldButtons];
+  return [...quickButtons, { type: "divider" }, ...fieldButtons];
 };
 
 export default async function ({ say, client, studentID, timestamp }) {
@@ -32,5 +32,5 @@ export default async function ({ say, client, studentID, timestamp }) {
     studentID,
     timestamp,
   });
-  await say({ blocks: homeBlocks, text: '', thread_ts: timestamp });
+  await say({ blocks: homeBlocks, text: "", thread_ts: timestamp });
 }
